replaced IndexingServiceClientMock with IIndexingServiceClientFallbackMock
